requirements.txt
black
cairosvg>=2.5.0
Pillow>=9.0.0
numpy-stl>=2.16.0
scikit-learn>=1.0.0
numpy>=1.20.0
scikit-image
cadquery
